integrate/fix download external content work into index.Rmd, especially ds 4 business; some notes on metaanalysis
@oskasf

Seems an OK approach?  Shall we carry this over to the other bookdowns where relevant?

From our chat:

[11:14 AM, 12/18/2020] David Reinstein: o let's make the 'real version' of `ds_for_bsns_notes.Rmd` the  `data_acad_materials/other_content_notes/ds_for_bsns_notes.Rmd` ... that's the one we edit.

then the index file `/metrics_discussion_work/index.Rmd` will pull down this file and rename it `...remote`  as in code above.

And never edit a file called 'remote'
and we can just choose a single 'real' version (in one repo) for each file and pull it in to other bookdowns/repos with the rename `..remote.Rmd`

 if someone only has access to the latter repo they can make a version suggesting changes called `..remote_edit.Rmd` and put in a pull request.